Rename error handler interface
